wip refactoring structure a bit
