+++ cmd/dumbp/cmd/markdown.go
@@ -1,4 +1,4 @@
-package commands
+package cmd
 
 import (
 	"errors"
@@ -7,9 +7,10 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/andyinabox/go-dumbphone/internal/utils"
 	"github.com/andyinabox/go-dumbphone/pkg/markdown"
+	"github.com/andyinabox/go-dumbphone/pkg/utils"
 	"github.com/manifoldco/promptui"
+	"github.com/spf13/viper"
 	"github.com/urfave/cli"
 )
 
@@ -97,24 +98,11 @@ var MarkdownSubcommand = cli.Command{
 
 		switch index {
 		case 0:
-			err := utils.BluetoothSend(file)
-			if err != nil {
-				return err
-			}
-			break
+			return utils.Send(file, utils.BLUETOOTH_SEND, "")
 		case 1:
-			return errors.New("USB transfer not yet implemented")
-			// err := utils.USBSend(file)
-			// if err != nil {
-			// 	return err
-			// }
-			break
+			return utils.Send(file, utils.USB_SEND, viper.GetString("modules.notes.dir"))
 		case 2:
-			err := utils.BrowserSend(file)
-			if err != nil {
-				return err
-			}
-			break
+			return utils.Send(file, utils.BROWSER_SEND, "")
 		}
 
 		return nil
