+++ cmd/dumbp/cmd/reader.go
@@ -1,12 +1,12 @@
-package commands
+package cmd
 
 import (
-	"errors"
 	"net/url"
 
-	"github.com/andyinabox/go-dumbphone/internal/utils"
 	"github.com/andyinabox/go-dumbphone/pkg/reader"
+	"github.com/andyinabox/go-dumbphone/pkg/utils"
 	"github.com/manifoldco/promptui"
+	"github.com/spf13/viper"
 	"github.com/urfave/cli"
 )
 
@@ -80,24 +80,11 @@ var ReaderSubcommand = cli.Command{
 
 		switch index {
 		case 0:
-			err := utils.BluetoothSend(file)
-			if err != nil {
-				return err
-			}
-			break
+			return utils.Send(file, utils.BLUETOOTH_SEND, "")
 		case 1:
-			return errors.New("USB transfer not yet implemented")
-			// err := utils.USBSend(file)
-			// if err != nil {
-			// 	return err
-			// }
-			break
+			return utils.Send(file, utils.USB_SEND, viper.GetString("modules.reader.dir"))
 		case 2:
-			err := utils.BrowserSend(file)
-			if err != nil {
-				return err
-			}
-			break
+			return utils.Send(file, utils.BROWSER_SEND, "")
 		}
 
 		return nil
