+++ pkg/usbsync/usbsync.go
@@ -1,0 +1,10 @@
+package usbsync
+
+import (
+	"errors"
+	"os"
+)
+
+func Send(f *os.File, path string) error {
+	return errors.New("File send not implemented yet")
+}
