+++ pkg/utils/timestamp.go
@@ -1,0 +1,11 @@
+package utils
+
+import (
+	"strconv"
+	"time"
+)
+
+// Timestamp creates a Unix-formatted string timestamp
+func Timestamp() string {
+	return strconv.FormatInt(time.Now().Unix(), 10)
+}
